+++ metadata.rb
@@ -1,7 +1,8 @@
 name             'jboss-eap'
-maintainer       'YOUR_COMPANY_NAME'
-maintainer_email 'YOUR_EMAIL'
+maintainer       'Andy Fraley'
+maintainer_email 'andrew.fraley@gmail.com'
 license          'All rights reserved'
 description      'Installs/Configures jboss-eap'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '1.0.0'
+version          '2.0.0'
+depends 'ark'